MuchThread#timeout
This implements the main `timeout` method.  This method takes a
seconds value and optional custom exception and a block.  It will
raise the custom exception (or `MuchTimeout::TimeoutError`if no
custom exception given) if calling the block takes more than
the given amount of seconds.

A few notes on this:

* this uses `IO.select` on an internal pipe to get accurate
  timeouts and to avoid having to `sleep`
* the block runs in a seperate thread from the main thread.  This
  is so we can safely raise in the block thread to interrupt and
  not affect the main thread.
* `timeout` requires a non-nil, numeric seconds value.  The idea
  here is that we want to protect from unintended side-effects or
  strange exceptions if giving `IO.select` non standard timeout
  values.
* `rescue false` on any pipe interactions to prevent errors from
  edge-case pipe scenarios from leaking unintended exceptions
* ensure we close any pipes to prevent any side-effects from
  unintentionally leaving file descriptors open

This is the base logic that the rest of the API will be built on.
In a coming effort, I'll implement the rest of the API.