build out the remaining api for the initial design
This adds the 3 additional api methods: `optional_timeout,
`just_timeout` and `just_optional_timeout`.  These methods each
provide special behavior in addition to the base `timeout` logic.

`optional_timeout` allows passing `nil` as the seconds value to
disable any timeout.  This is a thin wrapper to `timeout` that
saves the developer a manual `if secs.nil?` conditional.

`just_timeout` does the same logic as `timeout` except that it
only interrupts the block and never raises an exception.  Its
signature is different: you pass the block as the value of the
`:do` key in hash args.  You can optionally pass an `:on_timeout`
block to run custom logic when timeouts occur.

`just_optional_timeout` combines the signature and behavior of
`just_timeout` with the `nil` seconds behavior of `optional_timeout`.
It behaves as you would expect.

Note: I chose to not stub the nested api calls and instead layer
on (ie repeat) the tests of the nested calls.  I chose this to
not only *actually* test the intended behavior of each method,
but also to not bind the tests to the implementation.

This completes the initial design implementation.